move logic to storage/snapshots
