+++ tests/storage/snapshots/test_snapshots.py
@@ -18,12 +18,14 @@ from tests.storage.snapshots.constants import (
     WINDOWS_DIRECTORY_PATH,
 )
 from tests.storage.snapshots.utils import (
+    assert_restored_dv_pvc_predictable_names,
     expected_output_after_restore,
     fail_to_create_snapshot_no_permissions,
     start_windows_vm_after_restore,
+    vm_restore_with_prefix_policy,
 )
 from tests.storage.utils import assert_windows_directory_existence
-from utilities.constants import LS_COMMAND, TIMEOUT_1MIN, TIMEOUT_10SEC
+from utilities.constants import LS_COMMAND, TIMEOUT_1MIN, TIMEOUT_10MIN, TIMEOUT_10SEC
 from utilities.storage import run_command_on_vm_and_check_output
 from utilities.virt import restart_vm_wait_for_running_vm, running_vm
 
@@ -263,6 +265,48 @@ class TestRestoreSnapshots:
                     command=LS_COMMAND,
                     expected_result=expected_output_after_restore(1),
                 )
+
+    @pytest.mark.parametrize(
+        "rhel_vm_name, snapshot_with_content",
+        [
+            pytest.param(
+                {"vm_name": "vm-cnv-80304"},
+                {"number_of_snapshots": 1, "online_vm": False},
+                marks=pytest.mark.polarion("CNV-80304"),
+            ),
+        ],
+        indirect=True,
+    )
+    def test_restore_snapshot_with_predictable_names(
+        self,
+        admin_client,
+        rhel_vm_for_snapshot,
+        snapshot_with_content,
+    ):
+        if rhel_vm_for_snapshot.ready:
+            rhel_vm_for_snapshot.stop(wait=True)
+
+        source_volume_name = rhel_vm_for_snapshot.instance.spec.template.spec.volumes[0].name
+        restored_vm_name = f"{rhel_vm_for_snapshot.name}-restored"
+
+        with vm_restore_with_prefix_policy(
+            name=f"{restored_vm_name}-restore",
+            namespace=rhel_vm_for_snapshot.namespace,
+            vm_name=restored_vm_name,
+            snapshot_name=snapshot_with_content[0].name,
+            client=admin_client,
+            prefix_policy="PrefixTargetName",
+        ) as vm_restore:
+            vm_restore.wait_restore_done(timeout=TIMEOUT_10MIN)
+
+            restore_status = vm_restore.instance.status
+            assert_restored_dv_pvc_predictable_names(
+                restored_vm_name=restored_vm_name,
+                source_volume_name=source_volume_name,
+                restored_dv_name=restore_status.restores[0].dataVolumeName,
+                restored_pvc_name=restore_status.restores[0].persistentVolumeClaim,
+                volume_restore_policy=vm_restore.instance.spec.get("volumeRestorePolicy"),
+            )
 
 
 @pytest.mark.parametrize(
